test: Add tests to `Cartao` component.
- Added tests to cover render of the component.
- Refactored Cartao's structure.
